+++ pmcprg/copulas/_base.py
@@ -125,6 +125,8 @@ class CopulaEnum(CopulaDataMixin, Enum):
     PLACKETT         = 17, "Plackett", "Plackett",                "CopulaPlackett", True,  ["tau_k"],          [EPS_MINUS_ONE, ONE_MINUS_EPS], "pmcprg.copulas.explicit.plackett"
     GALAMBOS         = 18, "Galambos", "Galambos",                "CopulaGalambos", True,  ["tau_k"],          [0.0 + EPS, 1.0],              "pmcprg.copulas.extreme_value.galambos"
     HUSLER_REISS     = 19, "HuslerReiss", "Hüsler-Reiss",          "CopulaHuslerReiss", True, ["tau_k"],        [0.0 + EPS, 1.0],              "pmcprg.copulas.extreme_value.husler_reiss"
+    CLAYTON90        = 20, "Clayton90", "Clayton (90° rotation)",  "CopulaClayton90",   True, ["tau_k"],        [-1.0, 0.0 - EPS],             "pmcprg.copulas.archimedean.rotated"
+    CLAYTON270       = 21, "Clayton270", "Clayton (270° rotation)", "CopulaClayton270", True, ["tau_k"],        [-1.0, 0.0 - EPS],             "pmcprg.copulas.archimedean.rotated"
 
     def describe(self):
         return self.name, self.value